split out http runnable options to it's own file, add tests and single config helper
