+++ runnables/httpserver/runner.go
@@ -15,46 +15,8 @@ import (
 	"github.com/robbyt/go-supervisor/supervisor"
 )
 
-// Option represents a functional option for configuring Runner.
-type Option func(*Runner)
-
-// WithLogHandler sets a custom slog handler for the Runner instance.
-// For example, to use a custom JSON handler with debug level:
-//
-//	handler := slog.NewJSONHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelDebug})
-//	runner, err := httpserver.NewRunner(ctx, httpserver.WithConfigCallback(configCallback), httpserver.WithLogHandler(handler))
-func WithLogHandler(handler slog.Handler) Option {
-	return func(r *Runner) {
-		if handler != nil {
-			r.logger = slog.New(handler.WithGroup("httpserver.Runner"))
-		}
-	}
-}
-
-// WithContext sets a custom context for the Runner instance.
-// This allows for more granular control over cancellation and timeouts.
-func WithContext(ctx context.Context) Option {
-	return func(r *Runner) {
-		if ctx != nil {
-			r.ctx, r.cancel = context.WithCancel(ctx)
-		}
-	}
-}
-
-// WithConfigCallback sets the function that will be called to load or reload configuration.
-// This option is required when creating a new Runner.
-func WithConfigCallback(callback func() (*Config, error)) Option {
-	return func(r *Runner) {
-		r.configCallback = callback
-	}
-}
-
-// WithName sets the name of the Runner instance.
-func WithName(name string) Option {
-	return func(r *Runner) {
-		r.name = name
-	}
-}
+// ConfigCallback is the function type signature for the callback used to load initial config, and new config during Reload()
+type ConfigCallback func() (*Config, error)
 
 // Runner implements a configurable HTTP server that supports graceful shutdown,
 // dynamic reconfiguration, and state monitoring. It meets the Runnable, Reloadable,
@@ -62,7 +24,7 @@ func WithName(name string) Option {
 type Runner struct {
 	name           string
 	config         atomic.Pointer[Config]
-	configCallback func() (*Config, error)
+	configCallback ConfigCallback
 	bootLock       sync.Mutex
 	server         *http.Server
 	serverRunning  atomic.Bool
